Switch to non-chunk-based parsing for now
The current implementation of the external scanner does not seem to work
correctly when used with `.parse_with`. I don't see anything in the
tree-sitter documentation about this, so for now the easiest thing
is just to use `.parse` and pre-concatenate the rope chunks.

I suspect that the scanner might receive calls to serialize and
deserialize at chunk boundaries (although the docs do not specify
that this is the case). If so, we could redesign the scanner so that
parsing context is preserved across these calls, solving the problem.